+++ data/src/main/java/io/parrotsoftware/qa_data/datasources/impl/UserRemoteDataSourceImpl.kt
@@ -1,10 +1,8 @@
 package io.parrotsoftware.qa_data.datasources.impl
 
-import io.parrotsoftware.qa_data.domain.Credentials
-import io.parrotsoftware.qa_data.domain.RepositoryResult
-import io.parrotsoftware.qa_data.domain.Store
-import io.parrotsoftware.qa_data.managers.UserManager
 import io.parrotsoftware.qa_data.datasources.UserRemoteDataSource
+import io.parrotsoftware.qa_data.domain.RepositoryResult
+import io.parrotsoftware.qa_data.managers.UserManager
 import io.parrotsoftware.qa_network.domain.requests.ApiAuthRequest
 import io.parrotsoftware.qa_network.interactors.NetworkInteractor
 import io.parrotsoftware.qa_network.services.ParrotApi
@@ -52,23 +50,4 @@ class UserRemoteDataSourceImpl(
         return RepositoryResult()
     }
 
-    override suspend fun userExists(): RepositoryResult<Boolean> {
-        return RepositoryResult(userManager.isAuth())
-    }
-
-    override suspend fun getCredentials(): RepositoryResult<Credentials> {
-        return RepositoryResult(
-            Credentials(
-            userManager.getAccess(), userManager.getRefresh()
-        )
-        )
-    }
-
-    override suspend fun getStore(): RepositoryResult<Store> {
-        return RepositoryResult(
-            Store(
-            userManager.getStoreUuid(), userManager.getStoreName()
-        )
-        )
-    }
 }
