+++ data/src/main/java/io/parrotsoftware/qa_data/repositories/UserRepository.kt
@@ -1,5 +1,6 @@
 package io.parrotsoftware.qa_data.repositories
 
+import io.parrotsoftware.qa_data.datasources.UserLocalDataSource
 import io.parrotsoftware.qa_data.domain.Credentials
 import io.parrotsoftware.qa_data.domain.RepositoryResult
 import io.parrotsoftware.qa_data.domain.Store
@@ -10,30 +11,20 @@ import javax.inject.Inject
 class UserRepository
 @Inject constructor(
     private val userRemoteDataSource: UserRemoteDataSource,
-    private val userManagerD: UserManager
-    ) {
+    private val userLocalDataSource: UserLocalDataSource
+) {
 
     suspend fun userExists(): RepositoryResult<Boolean> =
-        userRemoteDataSource.userExists()
+        userLocalDataSource.userExists()
 
     suspend fun login(email: String, password: String): RepositoryResult<Nothing> =
         userRemoteDataSource.login(email, password)
 
-    suspend fun getCredentials(): RepositoryResult<Credentials> {
-        return RepositoryResult(
-            Credentials(
-            userManagerD.getAccess(), userManagerD.getRefresh()
-        )
-        )
-    }
+    suspend fun getCredentials(): RepositoryResult<Credentials> =
+        userLocalDataSource.getCredentials()
 
-     suspend fun getStore(): RepositoryResult<Store> {
-        return RepositoryResult(
-            Store(
-            userManagerD.getStoreUuid(), userManagerD.getStoreName()
-        )
-        )
-    }
+
+    suspend fun getStore(): RepositoryResult<Store> = userLocalDataSource.getStore()
 
 
 }
